+++ dms-20250414/alibabacloud_dms20250414/models/_create_data_agent_session_response_body.py
@@ -172,6 +172,7 @@ class CreateDataAgentSessionResponseBodyDataSessionConfig(DaraModel):
         language: str = None,
         mcp_server_ids: List[str] = None,
         mode: str = None,
+        user_oss_bucket: str = None,
     ):
         self.custom_agent_id = custom_agent_id
         self.custom_agent_stage = custom_agent_stage
@@ -179,6 +180,7 @@ class CreateDataAgentSessionResponseBodyDataSessionConfig(DaraModel):
         self.language = language
         self.mcp_server_ids = mcp_server_ids
         self.mode = mode
+        self.user_oss_bucket = user_oss_bucket
 
     def validate(self):
         pass
@@ -206,6 +208,9 @@ class CreateDataAgentSessionResponseBodyDataSessionConfig(DaraModel):
         if self.mode is not None:
             result['Mode'] = self.mode
 
+        if self.user_oss_bucket is not None:
+            result['UserOssBucket'] = self.user_oss_bucket
+
         return result
 
     def from_map(self, m: dict = None):
@@ -227,6 +232,9 @@ class CreateDataAgentSessionResponseBodyDataSessionConfig(DaraModel):
 
         if m.get('Mode') is not None:
             self.mode = m.get('Mode')
+
+        if m.get('UserOssBucket') is not None:
+            self.user_oss_bucket = m.get('UserOssBucket')
 
         return self